+++ src/ivec2.c
@@ -82,13 +82,13 @@ glmc_ivec2_distance(ivec2 a, ivec2 b) {
 
 CGLM_EXPORT
 void
-glmc_ivec2_fill(ivec2 v, float val) {
+glmc_ivec2_fill(ivec2 v, int val) {
   glm_ivec2_fill(v, val);
 }
 
 CGLM_EXPORT
 bool
-glmc_ivec2_eq(ivec2 v, float val) {
+glmc_ivec2_eq(ivec2 v, int val) {
   return glm_ivec2_eq(v, val);
 }
