+++ src/ivec3.c
@@ -82,13 +82,13 @@ glmc_ivec3_distance(ivec3 a, ivec3 b) {
 
 CGLM_EXPORT
 void
-glmc_ivec3_fill(ivec3 v, float val) {
+glmc_ivec3_fill(ivec3 v, int val) {
   glm_ivec3_fill(v, val);
 }
 
 CGLM_EXPORT
 bool
-glmc_ivec3_eq(ivec3 v, float val) {
+glmc_ivec3_eq(ivec3 v, int val) {
   return glm_ivec3_eq(v, val);
 }
